Tests/LibWeb/Text/expected/abortsignal-abort.txt

Aborted: true
Reason: "[object DOMException]"
Aborted: true
Reason: "This is a test"
